+++ src/client/components/PaginationPanel/PaginationPanel.react.js
@@ -4,7 +4,6 @@ import Pagination from 'react-bootstrap/lib/Pagination';
 import FormControl from 'react-bootstrap/lib/FormControl';
 import Button from 'react-bootstrap/lib/Button';
 import translations from './i18n';
-// import './PaginationPanel.less';
 
 const NEXT_BUTTON = (<span className="glyphicon glyphicon-forward" />);
 const PREV_BUTTON = (<span className="glyphicon glyphicon-backward" />);
@@ -80,49 +79,48 @@ export default class PaginationPanel extends Component {
   };
 
   render() {
-    const count = this.props.count;
-    const max = this.props.max;
+    const { count, max, offset } = this.props;
 
-    if (count > 0 && count > max) {
-      const offset = this.props.offset;
-      const items = Math.ceil(count / max);
-      const activePage = offset > 0 ? (offset / max + 1) : 1;
-
-      return (
-        <div>
-          <div className="pull-left">
-            <Pagination
-              items={items}
-              activePage={activePage}
-              maxButtons={3}
-              onSelect={this.onSelect.bind(this)}
-              prev={PREV_BUTTON}
-              next={NEXT_BUTTON}
-              first={false}
-              last={false}
-              ellipsis={true}
-              boundaryLinks={true}
-              style={{ verticalAlign: 'middle' }}
-            />
-          </div>
-          <div className="pull-left">
-            <FormControl
-              type="text"
-              className="form-control"
-              style={{ width: '50px', textAlign: 'center' }}
-              onChange={(e) => { this.setState({ pageNumber: e.target.value }) }}
-              onKeyPress={this.handleKeyPress}
-              value={this.state.pageNumber}
-            />
-          </div>
-          <div className="pull-left">
-            <Button className="btn btn-default" onClick={() => { this.onSelect(this.state.pageNumber) }}>
-              {this.context.i18n.getMessage('PaginationPanel.goToPage')}
-            </Button>
-          </div>
-        </div>
-      );
+    if (max === -1 || count === 0 || count <= max) {
+      return null
     }
-    return null
+
+    const items = Math.ceil(count / max);
+    const activePage = offset > 0 ? (offset / max + 1) : 1;
+
+    return (
+      <div>
+        <div className="pull-left">
+          <Pagination
+            items={items}
+            activePage={activePage}
+            maxButtons={3}
+            onSelect={this.onSelect.bind(this)}
+            prev={PREV_BUTTON}
+            next={NEXT_BUTTON}
+            first={false}
+            last={false}
+            ellipsis={true}
+            boundaryLinks={true}
+            style={{ verticalAlign: 'middle' }}
+          />
+        </div>
+        <div className="pull-left">
+          <FormControl
+            type="text"
+            className="form-control"
+            style={{ width: '50px', textAlign: 'center' }}
+            onChange={(e) => { this.setState({ pageNumber: e.target.value }) }}
+            onKeyPress={this.handleKeyPress}
+            value={this.state.pageNumber}
+          />
+        </div>
+        <div className="pull-left">
+          <Button className="btn btn-default" onClick={() => { this.onSelect(this.state.pageNumber) }}>
+            {this.context.i18n.getMessage('PaginationPanel.goToPage')}
+          </Button>
+        </div>
+      </div>
+    );
   }
 }